third_party/blink/web_tests/fast/hidpi/image-srcset-change-dynamically-from-js-1x-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.isPreloaded("resources/green-400-px-square.png") is false
PASS document.getElementById("foo").clientWidth==100 is true
This test passes if this img tag below is a blue square when the scale factor is 1. It ensures that attributes can be changed dynamically from javascript, and that only the correct resource is loaded.
